Reveal only the URL at the frame open time in performance.measureMemory
This changes the reporting in performance.measureMemory to reveal only
the URL at the frame creation time for cross-origin iframes and
thus prevents URL information leak if the iframe navigates after loading.

This is done by saving the first cross-origin URL loaded in LocalFrame
in LocalFrame::DidAttachDocument and using that URL for reporting.

The patch also adds a new test for redirecting iframe and fixes
the existing tests to wait until child iframes are loaded.

Bug:1049093
Change-Id: I0162efb82061f036450586dea0b191cacdce23f2