Add WPT for DedicatedWorker + COEP reporting
This CL adds tests to check the support of COEP reporting for requests
initiated by a dedicated worker. So far, Chrome has a bug that the
context url is the page's url where the dedicated is created. This bug
is going to be fixed once PlzDedicatedWorker is shipped (with additional
work for integration between PlzDedicatedWorker and COEP reporting).

Bug: 1060837
Change-Id: I12291646c4d5a80d89c8894548acdb186f4a14be